+++ CHANGELOG.md
@@ -13,6 +13,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Place entries relative to the currently selected entry when performing a move operation.
 - Add a command for loading random save files
 
+### Changed
+
+- Breaking: Rename the `set_savefile` configuration command to `set_save_file`.
+
 ### Fixed
 
 - Handle file move events where the destination or the source is outside the watched directories.